Fix zix_digest64() to consume all input
This was a copy-paste bug since the loop in zix_digest32() worked differently.
As a result only the first block was considered, making the digest nearly
useless for larger values.  The tests didn't (and unfortunately still don't)
catch this because the 64-bit digest algorithm incorporates the size itself.

Fix this by changing the loop to work the same way as zix_digest32(), so
hopefully something like this doesn't happen again.