Simplify ring space calculations
It isn't necessary to branch here to avoid underflow.  Essentially, the way
that unsigned binary integers work "automatically" does what this code was
doing.

Note that these expressions only work for a ring buffer that, like this one,
has a power of 2 (real) size and a maximum capacity 1 less than that.