Split up platform sources
This puts more onus on the build system to do things properly, but it's still
easy enough to build, even manually: all the files in the appropriate system
subdirectory just need to be included in the build.

Otherwise, the several nested levels of preprocessor conditionals get
confusing, and clang-format doesn't format code properly.