Skip socket file type test with absurdly long TMPDIR
If TMPDIR is so long that it doesn't fit in the socket address field, then it's
impossible to create a socket in the temporary directory.  This is a pretty
weird situation that would never happen on a normal system, so I think the test
can just be skipped without the reduced coverage being a problem.